Noiseprint/requirements.txt
astor==0.8.1
certifi==2021.5.30
charset-normalizer==2.0.7
cycler==0.10.0
decorator==4.4.2
grpcio==1.12.1
idna==3.3
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.15.0
opencv-python==4.2.0.34
pandas==0.25.3
Pillow==8.0.1
protobuf==3.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.3
PyWavelets==1.1.1
requests==2.26.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.5.4
six==1.15.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
tqdm==4.46.1
urllib3==1.26.7
wincertstore==0.2

